backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
pandas==2.2.2
python-dotenv==1.0.1
Requests==2.32.3
serpapi==0.1.5
tensorflow==2.12.0
yfinance==0.2.41
yahooquery==2.3.7
beautifulsoup4==4.12.3
pandas_ta==0.3.14b0